fix #327 0.2.1 build error, bump version to 0.2.2
