+++ conf/kafka.conf
@@ -25,7 +25,7 @@ kafka {
   grouper.factory.class = "org.apache.gearpump.streaming.kafka.lib.grouper.KafkaDefaultGrouperFactory"
 
   task {
-    message.decoder.class = "org.apache.gearpump.streaming.examples.kafka.util.KafkaDecoder"
-    timestamp.filter.class = "org.apache.gearpump.streaming.examples.kafka.util.KafkaFilter"
+    message.decoder.class = "org.apache.gearpump.streaming.kafka.lib.KafkaDecoder"
+    timestamp.filter.class = "org.apache.gearpump.streaming.kafka.lib.KafkaFilter"
   }
 }
