+++ external/kafka/src/main/scala/org/apache/gearpump/streaming/kafka/lib/KafkaFilter.scala
@@ -16,19 +16,13 @@
  * limitations under the License.
  */
 
-package org.apache.gearpump.streaming.examples.kafka.util
+package org.apache.gearpump.streaming.kafka.lib
 
-import com.twitter.bijection.Injection
-import org.apache.gearpump.Message
-import org.apache.gearpump.streaming.transaction.api.MessageDecoder
+import org.apache.gearpump.streaming.transaction.api.TimeStampFilter
+import org.apache.gearpump.{Message, TimeStamp}
 
-import scala.util.{Failure, Success}
-
-class KafkaDecoder extends MessageDecoder {
-  override def fromBytes(bytes: Array[Byte]): Message = {
-    Injection.invert[String, Array[Byte]](bytes) match {
-      case Success(s) => Message(s, System.currentTimeMillis())
-      case Failure(e) => throw e
-    }
+class KafkaFilter extends TimeStampFilter {
+  override def filter(msg: Message, predicate: TimeStamp): Option[Message] = {
+    Option(msg).find(_.timestamp >= predicate)
   }
 }
